# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Tracker for Products, Recipes, and Orders",
    "version" : "Layout",
    "title" : "Elevated Inventory Tracker",
    "contact" : {
      "email" : "joshydenberg@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/yjosh/ElevatedV3/1.0.0",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/products/{productId}" : {
      "get" : {
        "tags" : [ "employees", "technical" ],
        "summary" : "gets a list of products",
        "description" : "By passing in the appropriate options, you can search for\navailable inventory in the system\n",
        "operationId" : "getProduct",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "description" : "Id of the Product you are looking for",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip for pagination",
          "required" : false,
          "type" : "integer",
          "minimum" : 0,
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "maximum number of records to return",
          "required" : false,
          "type" : "integer",
          "maximum" : 100,
          "minimum" : 0,
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Product object",
            "schema" : {
              "$ref" : "#/definitions/Product"
            }
          },
          "400" : {
            "description" : "Unable to find products"
          }
        }
      },
      "put" : {
        "tags" : [ "employees", "technical" ],
        "summary" : "Edit a specific product",
        "description" : "Edit a specific product",
        "operationId" : "editProduct",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "description" : "Id of the Product you are looking for",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip for pagination",
          "required" : false,
          "type" : "integer",
          "minimum" : 0,
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "maximum number of records to return",
          "required" : false,
          "type" : "integer",
          "maximum" : 100,
          "minimum" : 0,
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "product",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Product"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Product object",
            "schema" : {
              "$ref" : "#/definitions/Product"
            }
          },
          "400" : {
            "description" : "Unable to find products"
          }
        }
      },
      "delete" : {
        "tags" : [ "admins", "employees", "manager", "technical" ],
        "summary" : "delete a product",
        "description" : "Adds a product to the system'",
        "operationId" : "deleteProduct",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "description" : "Id of the Product you are looking for",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Product object that was deleted",
            "schema" : {
              "$ref" : "#/definitions/Product"
            }
          },
          "400" : {
            "description" : "invalid input, object invalid"
          },
          "409" : {
            "description" : "unable to delete object"
          }
        }
      }
    },
    "/products/{productId}/recipes" : {
      "get" : {
        "tags" : [ "employees", "technical" ],
        "summary" : "Retreives the recipe for a specific product",
        "description" : "Retreives the recipe for a specific product By passing in the appropriate options, you can search for available inventory in the system",
        "operationId" : "searchRecipeByProduct",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "description" : "Id of the Product you are looking for",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip for pagination",
          "required" : false,
          "type" : "integer",
          "minimum" : 0,
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "maximum number of records to return",
          "required" : false,
          "type" : "integer",
          "maximum" : 100,
          "minimum" : 0,
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Found Valid list of products",
            "schema" : {
              "$ref" : "#/definitions/Recipe"
            }
          },
          "400" : {
            "description" : "Unable to find valid product"
          }
        }
      },
      "put" : {
        "tags" : [ "employees", "technical" ],
        "summary" : "Retreives the recipe for a specific product",
        "description" : "Retreives the recipe for a specific product By passing in the appropriate options, you can search for available inventory in the system",
        "operationId" : "editRecipeByProduct",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "description" : "Id of the Product you are looking for",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip for pagination",
          "required" : false,
          "type" : "integer",
          "minimum" : 0,
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "maximum number of records to return",
          "required" : false,
          "type" : "integer",
          "maximum" : 100,
          "minimum" : 0,
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Edited Recipe for productId",
            "schema" : {
              "type" : "object",
              "properties" : {
                "productId" : {
                  "type" : "string"
                },
                "recipe" : {
                  "$ref" : "#/definitions/Recipe"
                }
              }
            }
          },
          "400" : {
            "description" : "Unable to find valid product"
          }
        }
      },
      "delete" : {
        "tags" : [ "admins", "technical" ],
        "summary" : "Removes a Product from a recipe",
        "description" : "Deletes an item to the system",
        "operationId" : "DeleteProduct",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "productId",
          "in" : "path",
          "description" : "Id of the Product you are looking for",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "productId",
          "in" : "query",
          "description" : "Id of the product to Delete",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Deleted Recipe from Product",
            "schema" : {
              "type" : "object",
              "properties" : {
                "recipeId" : {
                  "type" : "string"
                },
                "productId" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "description" : "invalid input, object does not exist"
          }
        }
      }
    },
...
    "/materials/{materialId}" : {
      "get" : {
        "tags" : [ "employees", "technical" ],
        "summary" : "Retreives a List of recipes, or a specific recipe",
        "description" : "Search for a specific recipe By passing in the appropriate options, you can search for available inventory in the system",
        "operationId" : "getMaterial",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "materialId",
          "in" : "path",
          "description" : "Id of the Recipe you are looking for",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip for pagination",
          "required" : false,
          "type" : "integer",
          "minimum" : 0,
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "maximum number of records to return",
          "required" : false,
          "type" : "integer",
          "maximum" : 100,
          "minimum" : 0,
          "format" : "int32"
        }, {
          "name" : "recipeId",
          "in" : "query",
          "description" : "maximum number of records to return",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Found a Valid material"
          },
          "400" : {
            "description" : "Unable to find valid material"
          }
        }
      },
      "put" : {
        "tags" : [ "employees", "technical" ],
        "summary" : "Edit a specific material",
        "description" : "Edit a specific material",
        "operationId" : "editMaterial",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "materialId",
          "in" : "path",
          "description" : "Id of the Recipe you are looking for",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip for pagination",
          "required" : false,
          "type" : "integer",
          "minimum" : 0,
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "maximum number of records to return",
          "required" : false,
          "type" : "integer",
          "maximum" : 100,
          "minimum" : 0,
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "material",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Material"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Material object",
            "schema" : {
              "$ref" : "#/definitions/Material"
            }
          },
          "400" : {
            "description" : "Unable to find materials"
          }
        }
      },
      "delete" : {
        "tags" : [ "admins", "technical" ],
        "summary" : "deletes an inventory item",
        "description" : "Deletes an item to the system",
        "operationId" : "deleteMaterial",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "materialId",
          "in" : "path",
          "description" : "Id of the Recipe you are looking for",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "201" : {
            "description" : "item created"
          },
          "400" : {
            "description" : "invalid input, object does not exist"
          }
        }
      }
    },
...
  },
  "definitions" : {
    "Product" : {
      "type" : "object",
      "required" : [ "category", "name", "productId", "unit" ],
      "properties" : {
        "productId" : {
          "type" : "string",
          "format" : "uuid",
          "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "recipeId" : {
          "type" : "string",
          "format" : "uuid",
          "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "name" : {
          "type" : "string",
          "example" : "Herbal Essences"
        },
        "sizes" : {
          "type" : "array",
          "xml" : {
            "name" : "size",
            "wrapped" : true
          },
          "items" : {
            "$ref" : "#/definitions/Size"
          }
        },
        "unit" : {
          "type" : "string",
          "example" : "2016-08-29T09:12:33.001Z"
        },
        "category" : {
          "type" : "string",
          "example" : "Shampoos"
        }
      }
    },
    "Size" : {
      "type" : "object",
      "properties" : {
        "size" : {
          "type" : "integer",
          "format" : "int64"
        },
        "unit" : {
          "type" : "string"
        }
      }
    },
    "Recipe" : {
      "type" : "object",
      "required" : [ "recipeId", "unit", "value" ],
      "properties" : {
        "productId" : {
          "type" : "string",
          "format" : "uuid",
          "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "recipeId" : {
          "type" : "string",
          "format" : "uuid",
          "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "materials" : {
          "type" : "object",
          "properties" : { }
        },
        "unit" : {
          "type" : "string",
          "example" : "2016-08-29T09:12:33.001Z"
        },
        "value" : {
          "type" : "number"
        }
      }
    },
...
    "Material" : {
      "type" : "object",
      "properties" : {
        "materialId" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "stockQuanitity" : {
          "type" : "number"
        },
        "unit" : {
          "type" : "string"
        },
        "supplierId" : {
          "type" : "string"
        }
      }
    },
...
  }
}